LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/choice-of-error-2-expected.txt

CONSOLE MESSAGE: SyntaxError: Importing binding name 'default' cannot be resolved by star export entries.
CONSOLE MESSAGE: SyntaxError: Importing binding name 'default' cannot be resolved by star export entries.

FAIL Instantiation errors in different files should be reported depending on different roots assert_not_equals: two different instantiation errors should be reported got disallowed value object "SyntaxError: Importing binding name 'default' cannot be resolved by star export entries."
